Var & const reformat
