+++ .github/workflows/deps.yml
@@ -2,8 +2,7 @@
 name: dependency-submission
 on:
   push:
-    branches:
-      - master
+    branches: [ main ]
 
 permissions:
   contents: write
