+++ internal/monitor/github/github_client.go
@@ -11,12 +11,13 @@ import (
 	"github.com/soltanoff/go_github_release_monitor_bot/pkg/logs"
 )
 
-var (
-	tagURIPattern  = regexp.MustCompile(`refs/tags/([\w\d\-\.]+)`)
-	releaseURLMask = "https://api.github.com/repos/%s/releases/latest"
-	tagsURLMask    = "https://api.github.com/repos/%s/git/refs/tags"
-	releaseTagMask = "https://github.com/%s/releases/tag/%s"
+const (
+	releaseURLMask string = "https://api.github.com/repos/%s/releases/latest"
+	tagsURLMask    string = "https://api.github.com/repos/%s/git/refs/tags"
+	releaseTagMask string = "https://github.com/%s/releases/tag/%s"
 )
+
+var tagURIPattern = regexp.MustCompile(`refs/tags/([\w\d\-\.]+)`)
 
 type ReleaseInfo struct {
 	TagName   string `json:"tag_name"`
